add AbstractStackOwnableEntity
