+++ src/main/java/at/minecraftschurli/mods/arsmagicalegacy/entity/NatureScythe.java
@@ -1,11 +1,7 @@
 package at.minecraftschurli.mods.arsmagicalegacy.entity;
 
-import at.minecraftschurli.mods.arsmagicalegacy.api.ArsMagicaApi;
 import at.minecraftschurli.mods.arsmagicalegacy.init.AMDamageTypes;
 import at.minecraftschurli.mods.arsmagicalegacy.util.AMUtil;
-import net.minecraft.network.syncher.EntityDataAccessor;
-import net.minecraft.network.syncher.EntityDataSerializers;
-import net.minecraft.network.syncher.SynchedEntityData;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.Entity;
@@ -13,39 +9,17 @@ import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.storage.ValueInput;
-import net.minecraft.world.level.storage.ValueOutput;
 import net.minecraft.world.phys.EntityHitResult;
 import net.minecraft.world.phys.HitResult;
 import net.neoforged.neoforge.entity.PartEntity;
 
-public class NatureScythe extends AbstractOwnableEntity {
-    private static final EntityDataAccessor<ItemStack> STACK = SynchedEntityData.defineId(NatureScythe.class, EntityDataSerializers.ITEM_STACK);
-    private static final String STACK_KEY = "stack";
+public class NatureScythe extends AbstractStackOwnableEntity {
     private boolean hasHit = false;
     private int hitTicks = -1;
 
     public NatureScythe(EntityType<? extends NatureScythe> type, Level level) {
         super(type, level);
-    }
-
-    @Override
-    protected void defineSynchedData(SynchedEntityData.Builder entityData) {
-        super.defineSynchedData(entityData);
-        entityData.define(STACK, ItemStack.EMPTY);
-    }
-
-    @Override
-    protected void readAdditionalSaveData(ValueInput input) {
-        input.child(ArsMagicaApi.MOD_ID).ifPresent(child -> entityData.set(STACK, child.read(STACK_KEY, ItemStack.CODEC).orElse(ItemStack.EMPTY)));
-    }
-
-    @Override
-    protected void addAdditionalSaveData(ValueOutput output) {
-        ValueOutput child = output.child(ArsMagicaApi.MOD_ID);
-        child.store(STACK_KEY, ItemStack.CODEC, entityData.get(STACK));
     }
 
     @Override
@@ -88,14 +62,6 @@ public class NatureScythe extends AbstractOwnableEntity {
             setHasHit();
         }
         setPos(position().add(getDeltaMovement()));
-    }
-
-    public ItemStack getStack() {
-        return entityData.get(STACK);
-    }
-
-    public void setStack(ItemStack stack) {
-        entityData.set(STACK, stack);
     }
 
     private void setHasHit() {
